refactor: Update User-Agent header in API requests
This commit updates the User-Agent header in the API requests made in the 0-subs.py and 1-top_ten.py files. The previous User-Agent value was "linux:0x16.api.advanced:v1.0.0 (by /u/bdov_)", which is specific to the developer. The new User-Agent value is "Mozilla/5.0", which is a more generic and widely accepted User-Agent. This change improves the compatibility and reliability of the API requests.