refactor: Update README with additional resources
This commit adds additional resources to the README.md file in the 0x17-web_stack_debugging_3 directory. The resources include links to Puppet-Lint and Puppet-CookBook, which provide helpful information and guides for using Puppet. This update improves the documentation and provides users with more information and references for working with Puppet.