refactor: Update Nginx configuration to serve /number_odd_or_even/ route
This commit updates the Nginx configuration to serve the /number_odd_or_even/ route by modifying the location block. The proxy_pass directive is set to http://127.0.0.1:5001/number_odd_or_even/$1 to correctly route the requests. This change aims to improve the handling of the number_odd_or_even route in the AirBnB application server.