refactor: Update gunicorn.service to configure working directory and syslog identifier
This commit updates the gunicorn.service file to configure the working directory and syslog identifier for the AirBnB application server. The WorkingDirectory directive is set to /home/ubuntu/AirBnB_clone_v4 to specify the correct directory for the server. Additionally, the SyslogIdentifier is set to "gunicorn" to identify the logs generated by Gunicorn. These changes aim to improve the organization and logging capabilities of the application server.